pre-testing timemanager, see desc
added additional methods to the treenode interface in order to establish a tree easier

this is a "hard" tree bc the connections are very literal. Note that this does not compile, and is just a coding approach that works in theory. If performance or "finding children" becomes an issue, try adapting to this variant instead.