work progress (see desc)
I need to redesign how pipes work in their entirety to have a two-way pipe implementation. Since any relationship between buildings has the potential to be in superposition (refinery to a hanging pipe could be anything, for instnace), the hard-coded parent-child fields no longer work.

I need to define the relationship in a way where buildings can have superpositions, outputs, or inputs. Maybe make a single list with tuples with Flowable-RelationshipState values.

That way, they still exist relative to each other in the tree, but can be iterated through when a constraint (e.g. a Well) is added. Tough beans, though. Ask if this is even necessary.